requirements.txt
pymongo==4.7.2
flask==3.0.3
flask_cors==5.0.0
python-dotenv==1.0.1
gunicorn
